fix: get_metadata actor fix
