+++ rust/candid_parser/src/typing.rs
@@ -1,19 +1,18 @@
 use crate::{parse_idl_prog, pretty_parse_idl_prog, Error, Result};
 use candid::types::{
     syntax::{
-        Binding, Dec, FuncType, IDLArgType, IDLEnv, IDLInitArgs, IDLProg, IDLType, PrimType,
-        TypeField,
+        Binding, Dec, FuncType, IDLArgType, IDLEnv, IDLInitArgs, IDLProg, IDLType, TypeField,
     },
-    ArgType, Field, Function, Type, TypeEnv, TypeInner,
+    ArgType, Type, TypeEnv, TypeInner,
 };
 use candid::utils::check_unique;
 use std::collections::BTreeMap;
 use std::path::{Path, PathBuf};
 
 pub struct Env<'a> {
-    pub te: &'a mut TypeEnv,
-    pub idl_env: &'a mut IDLEnv,
-    pub pre: bool,
+    te: &'a mut TypeEnv,
+    idl_env: &'a mut IDLEnv,
+    pre: bool,
 }
 
 impl Env<'_> {
@@ -24,130 +23,7 @@ impl Env<'_> {
 
 /// Convert candid AST to internal Type
 pub fn ast_to_type(env: &TypeEnv, ast: &IDLType) -> Result<Type> {
-    let env = Env {
-        te: &mut env.clone(),
-        idl_env: &mut IDLEnv::new(),
-        pre: false,
-    };
-    map_type(&env, ast)
-}
-
-fn map_prim(prim: &PrimType) -> Type {
-    match prim {
-        PrimType::Nat => TypeInner::Nat,
-        PrimType::Nat8 => TypeInner::Nat8,
-        PrimType::Nat16 => TypeInner::Nat16,
-        PrimType::Nat32 => TypeInner::Nat32,
-        PrimType::Nat64 => TypeInner::Nat64,
-        PrimType::Int => TypeInner::Int,
-        PrimType::Int8 => TypeInner::Int8,
-        PrimType::Int16 => TypeInner::Int16,
-        PrimType::Int32 => TypeInner::Int32,
-        PrimType::Int64 => TypeInner::Int64,
-        PrimType::Float32 => TypeInner::Float32,
-        PrimType::Float64 => TypeInner::Float64,
-        PrimType::Bool => TypeInner::Bool,
-        PrimType::Text => TypeInner::Text,
-        PrimType::Null => TypeInner::Null,
-        PrimType::Reserved => TypeInner::Reserved,
-        PrimType::Empty => TypeInner::Empty,
-    }
-    .into()
-}
-
-pub fn map_type(env: &Env, t: &IDLType) -> Result<Type> {
-    match t {
-        IDLType::PrimT(prim) => Ok(map_prim(prim)),
-        IDLType::VarT(id) => {
-            env.te.find_type(id)?;
-            Ok(TypeInner::Var(id.to_string()).into())
-        }
-        IDLType::OptT(t) => {
-            let t = map_type(env, t)?;
-            Ok(TypeInner::Opt(t).into())
-        }
-        IDLType::VecT(t) => {
-            let t = map_type(env, t)?;
-            Ok(TypeInner::Vec(t).into())
-        }
-        IDLType::RecordT(fs) => {
-            let fs = map_fields(env, fs)?;
-            Ok(TypeInner::Record(fs).into())
-        }
-        IDLType::VariantT(fs) => {
-            let fs = map_fields(env, fs)?;
-            Ok(TypeInner::Variant(fs).into())
-        }
-        IDLType::PrincipalT => Ok(TypeInner::Principal.into()),
-        IDLType::FuncT(func) => {
-            let mut t1 = Vec::new();
-            for arg in func.args.iter() {
-                t1.push(map_arg(env, arg)?);
-            }
-            let mut t2 = Vec::new();
-            for t in func.rets.iter() {
-                t2.push(map_type(env, t)?);
-            }
-            if func.modes.len() > 1 {
-                return Err(Error::msg("cannot have more than one mode"));
-            }
-            if func.modes.len() == 1
-                && func.modes[0] == candid::types::FuncMode::Oneway
-                && !t2.is_empty()
-            {
-                return Err(Error::msg("oneway function has non-unit return type"));
-            }
-            let f = Function {
-                modes: func.modes.clone(),
-                args: t1,
-                rets: t2,
-            };
-            Ok(TypeInner::Func(f).into())
-        }
-        IDLType::ServT(ms) => {
-            let ms = map_meths(env, ms)?;
-            Ok(TypeInner::Service(ms).into())
-        }
-        IDLType::ClassT(_, _) => Err(Error::msg("service constructor not supported")),
-        IDLType::UnknownT => Err(Error::msg("unknown type")),
-    }
-}
-
-fn map_arg(env: &Env, arg: &IDLArgType) -> Result<ArgType> {
-    Ok(ArgType {
-        name: arg.name.clone(),
-        typ: map_type(env, &arg.typ)?,
-    })
-}
-
-fn map_fields(env: &Env, fs: &[TypeField]) -> Result<Vec<Field>> {
-    // field label duplication is checked in the parser
-    let mut res = Vec::new();
-    for f in fs.iter() {
-        let ty = map_type(env, &f.typ)?;
-        let field = Field {
-            id: f.label.clone().into(),
-            ty,
-        };
-        res.push(field);
-    }
-    Ok(res)
-}
-
-fn map_meths(env: &Env, ms: &[Binding]) -> Result<Vec<(String, Type)>> {
-    // binding duplication is checked in the parser
-    let mut res = Vec::new();
-    for meth in ms.iter() {
-        let t = map_type(env, &meth.typ)?;
-        if !env.pre && env.te.as_func(&t).is_err() {
-            return Err(Error::msg(format!(
-                "method {} is a non-function type",
-                meth.id
-            )));
-        }
-        res.push((meth.id.to_owned(), t));
-    }
-    Ok(res)
+    env.map_type(ast).map_err(Error::msg)
 }
 
 pub fn check_type(env: &Env, t: &IDLType) -> Result<IDLType> {
@@ -396,13 +272,12 @@ fn merge_actor(
         ))),
         Some(idl_type) => {
             let t = env.idl_env.trace_type(idl_type).map_err(Error::msg)?;
-            let idl_meths = env.idl_env.as_service(idl_type).map_err(Error::msg)?;
             match &t {
                 IDLType::ClassT(_, _) => Err(Error::msg(format!(
                     "Imported service file {file:?} has a service constructor"
                 ))),
                 IDLType::ServT(meths) => match &actor {
-                    None => Ok(Some(t.clone())),
+                    None => Ok(Some(t)),
                     Some(idl_type) => {
                         let t = env.idl_env.trace_type(idl_type).map_err(Error::msg)?;
                         let serv = env.idl_env.as_service(&t).map_err(Error::msg)?;
@@ -411,8 +286,7 @@ fn merge_actor(
                         check_unique(ms.iter().map(|m| &m.id)).map_err(|e| {
                             Error::msg(format!("Duplicate imported method name: {e}"))
                         })?;
-                        let res =
-                            IDLType::ServT(idl_meths.iter().chain(meths.iter()).cloned().collect());
+                        let res = IDLType::ServT(ms);
                         Ok(Some(res))
                     }
                 },
